Moving export all activity management command to task
